OperatorTimeInterval
